MedLog/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=MedLog/backend/requirements.txt MedLog/backend/pyproject.toml
#
aiosqlite==0.21.0
    # via medlogserver (MedLog/backend/pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
apscheduler==3.11.0
    # via medlogserver (MedLog/backend/pyproject.toml)
authlib==1.5.0
    # via medlogserver (MedLog/backend/pyproject.toml)
bcrypt==4.2.1
    # via passlib
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.8
    # via
    #   medlogserver (MedLog/backend/pyproject.toml)
    #   pip-tools
    #   uvicorn
cryptography==44.0.1
    # via
    #   authlib
    #   python-jose
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via pydantic
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.8
    # via medlogserver (MedLog/backend/pyproject.toml)
getversion==1.0.2
    # via medlogserver (MedLog/backend/pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via medlogserver (MedLog/backend/pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
importlib-metadata==8.6.1
    # via build
itsdangerous==2.2.0
    # via medlogserver (MedLog/backend/pyproject.toml)
oauthlib==3.2.2
    # via medlogserver (MedLog/backend/pyproject.toml)
packaging==24.2
    # via
    #   build
    #   setuptools-scm
passlib[bcrypt]==1.7.4
    # via medlogserver (MedLog/backend/pyproject.toml)
pip-tools==7.4.1
    # via medlogserver (MedLog/backend/pyproject.toml)
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.10.6
    # via
    #   fastapi
    #   medlogserver (MedLog/backend/pyproject.toml)
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.0
    # via medlogserver (MedLog/backend/pyproject.toml)
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via pydantic-settings
python-jose[cryptography]==3.4.0
    # via medlogserver (MedLog/backend/pyproject.toml)
python-multipart==0.0.20
    # via medlogserver (MedLog/backend/pyproject.toml)
pyyaml==6.0.2
    # via medlogserver (MedLog/backend/pyproject.toml)
rsa==4.9
    # via python-jose
setuptools-scm==8.2.0
    # via getversion
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy[asyncio]==2.0.38
    # via
    #   medlogserver (MedLog/backend/pyproject.toml)
    #   sqlmodel
sqlmodel==0.0.22
    # via medlogserver (MedLog/backend/pyproject.toml)
starlette==0.45.3
    # via
    #   fastapi
    #   medlogserver (MedLog/backend/pyproject.toml)
stdlib-list==0.11.1
    # via getversion
tomli==2.2.1
    # via
    #   build
    #   pip-tools
    #   setuptools-scm
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   setuptools-scm
    #   sqlalchemy
    #   starlette
    #   uvicorn
tzlocal==5.3
    # via apscheduler
uvicorn==0.34.0
    # via medlogserver (MedLog/backend/pyproject.toml)
wheel==0.45.1
    # via pip-tools
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
